% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFun.R
\name{testFun}
\alias{testFun}
\title{Make a binary factor}
\usage{
testFun(n)
}
\arguments{
\item{n}{integer}
}
\value{
factor
}
\description{
Create a factor vector of n elements, labeled A and B.
}
\examples{
testFun(76)
}
